backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
google-generativeai
langchain==0.1.0
langchain-google-genai
langchain-core
pip-audit==2.6.1
typing-extensions==4.9.0
protobuf==4.25.9
tqdm==4.67.3
PyJWT
asyncpg
pydantic[email]
langgraph